ash-archive/pyproject.toml
[project]
name = "ash-archive-control"
version = "0.1.0"
description = "Control repository for Ash Archive dual-edition Morrowind modlist planning"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
